rework signify arguments for compatiblity with signify and minisign
Both signify and minisign have a concept of untrusted comment,
which if signing by default seems to point to the pubkey.
If available, the option for untrusted comment only ever seems to be -c.

Only minisign seems to have a concept of a trusted comment
and the option for a trusted comment is -t

This commit makes the trusted comment optional
(defaults to off, previously defaulted to time.asctime()).
This removes the unneeded signing and prompt round for those
who do not need or use the trusted comment section.

This commit also brings the `trezor-signify` command line closer
to the minisign and signify tools, which unfortunately seems to be
a breaking change, but the previous state seemed to be more messy.